Move and split vowl element visitors
