+++ src/main/java/de/uni_stuttgart/vis/vowl/owl2vowl/model/entities/nodes/classes/NullClass.java
@@ -1,6 +1,6 @@
 package de.uni_stuttgart.vis.vowl.owl2vowl.model.entities.nodes.classes;
 
-import de.uni_stuttgart.vis.vowl.owl2vowl.parser.vowl.VowlElementVisitor;
+import de.uni_stuttgart.vis.vowl.owl2vowl.model.visitor.VowlElementVisitor;
 
 /**
  * NullClass which has no functionality and should be just used for VisitorEx returns.
