+++ src/main/java/de/uni_stuttgart/vis/vowl/owl2vowl/model/individuals/VowlIndividual.java
@@ -6,7 +6,7 @@
 package de.uni_stuttgart.vis.vowl.owl2vowl.model.individuals;
 
 import de.uni_stuttgart.vis.vowl.owl2vowl.model.AbstractVowlObject;
-import de.uni_stuttgart.vis.vowl.owl2vowl.parser.vowl.VowlElementVisitor;
+import de.uni_stuttgart.vis.vowl.owl2vowl.model.visitor.VowlElementVisitor;
 import org.semanticweb.owlapi.model.IRI;
 
 /**
